refactor(tests): extract shared ElevenLabs helpers + 120s boundary test
Addresses claude bot review on PR #105: dedupe _el_config /
_make_elevenlabs_client_mock / FAKE_EL_MP3 into tests/helpers_elevenlabs.py
(also de-duplicating test_stems.py's copy) and add an exact-120s
single-chunk boundary test for keep-range splitting.
MIME-type suggestion rebutted: httpx infers content-type from the upload
filename; the identical pattern is proven against the live API by stems (#97).
