+++ tests/test_repaint.py
@@ -14,6 +14,7 @@ from typer.testing import CliRunner
 
 from acemusic.cli import app
 from acemusic.models import Clip
+from tests.helpers_elevenlabs import FAKE_EL_MP3, _el_config, _make_elevenlabs_client_mock
 
 runner = CliRunner()
 
@@ -343,31 +344,6 @@ class TestCrossfadeStitchHelper:
 # ---------------------------------------------------------------------------
 # ElevenLabs backend (#98)
 # ---------------------------------------------------------------------------
-
-FAKE_EL_MP3 = b"ID3" + b"\x00" * 200
-
-
-def _el_config(monkeypatch, api_key="test-key", output_format="mp3_44100_128"):
-    """Point load_config at an ElevenLabs-enabled config."""
-    from acemusic.config import AceConfig
-
-    monkeypatch.setattr(
-        "acemusic.cli.load_config",
-        lambda: AceConfig(
-            api_url="http://localhost:8001",
-            api_key=None,
-            elevenlabs_api_key=api_key,
-            elevenlabs_output_format=output_format,
-        ),
-    )
-
-
-def _make_elevenlabs_client_mock(audio_bytes: bytes = FAKE_EL_MP3, song_id: str = "song-123"):
-    """MagicMock ElevenLabsClient with a happy-path upload→plan→compose default."""
-    el = MagicMock()
-    el.upload_for_inpainting.return_value = song_id
-    el.generate_from_plan.return_value = audio_bytes
-    return el
 
 
 @pytest.fixture
